zadaniedjango/requirements.txt
Django==1.11.2
python-dateutil==2.6.0
